+++ lib/boruta/oauth/applications/authorize_application.ex
@@ -6,14 +6,14 @@ defmodule Boruta.Oauth.AuthorizeApplication do
   """
 
   @doc """
-  This function will be triggered in case of success triggering `Boruta.Oauth.preauthorize/3`
+  This function will be triggered in case of success invoking `Boruta.Oauth.preauthorize/3`
   """
   @callback preauthorize_success(
               conn :: Plug.Conn.t(),
               authorization :: %Boruta.Oauth.AuthorizationSuccess{}
             ) :: any()
   @doc """
-  This function will be triggered in case of failure triggering `Boruta.Oauth.preauthorize/3`
+  This function will be triggered in case of failure invoking `Boruta.Oauth.preauthorize/3`
   """
   @callback preauthorize_error(
               conn :: Plug.Conn.t(),
@@ -22,14 +22,14 @@ defmodule Boruta.Oauth.AuthorizeApplication do
               any()
 
   @doc """
-  This function will be triggered in case of success triggering `Boruta.Oauth.authorize/3`
+  This function will be triggered in case of success invoking `Boruta.Oauth.authorize/3`
   """
   @callback authorize_success(
               conn :: Plug.Conn.t(),
               authorize_response :: Boruta.Oauth.AuthorizeResponse.t()
             ) :: any()
   @doc """
-  This function will be triggered in case of failure triggering `Boruta.Oauth.authorize/3`
+  This function will be triggered in case of failure invoking `Boruta.Oauth.authorize/3`
   """
   @callback authorize_error(conn :: Plug.Conn.t(), oauth_error :: Boruta.Oauth.Error.t()) ::
               any()